% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectPatternType.R
\name{detectPatternType}
\alias{detectPatternType}
\title{Detect a type of pattern for each pair}
\usage{
detectPatternType(dmix.list, side.nqs = NULL, nq.list, isoformLevel.data,
  geneLevel.data = NULL)
}
\arguments{
\item{dmix.list}{A list of detected models}

\item{side.nqs}{A list of type of sides of models}

\item{nq.list}{A list of number of components of detected models}

\item{isoformLevel.data}{A data matrix of isoform expression}

\item{geneLevel.data}{A data matrix of gene expression}
}
\value{
A list consisting of types of patterns and merged types of patterns
}
\description{
Detect a type of pattern for each pair
}
\examples{
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("AnnotationDbi")
txdb=TxDb.Hsapiens.UCSC.hg19.knownGene
#The txdb also can be imported from a gtf file by using package GenomicFeatures
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=logC(isoformDataSample)
#now data is ready
tbreak=round(sqrt(ncol(isoformDataSample)))
model.res=doMixtureModelMatrix(isoformLevel.data=isoformDataSample,txdb=txdb,tbreak=tbreak)
res=detectPatternType(dmix.list=model.res$dmix.list,
    nq.list=model.res$nq.list,isoformLevel.data=isoformDataSample)
pie(table(res$pattern.type),col=colors()[c(12,11,132,131,137,136)])
}

