% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSide.R
\name{detectSide}
\alias{detectSide}
\title{Detect sides of models}
\usage{
detectSide(dmix.list, nq.vec)
}
\arguments{
\item{dmix.list}{A list of detected models}

\item{nq.vec}{A vector of number of components of detected models}
}
\value{
A vector consisting of sides of models
}
\description{
Detect sides of models
}
\examples{
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("AnnotationDbi")
txdb=TxDb.Hsapiens.UCSC.hg19.knownGene
#The txdb also can be imported from a gtf file by using package GenomicFeatures
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=ifelse(isoformDataSample==0,0,log2(isoformDataSample))
#now data is ready
tbreak=round(sqrt(ncol(isoformDataSample)))
model.res=doMixtureModelMatrix(isoformLevel.data=isoformDataSample,txdb=txdb,tbreak=tbreak)
res=detectSide(dmix.list=model.res$dmix.list,nq.vec=unlist(model.res$nq.list))
}
