% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doGaussianMixtureModel.R
\name{doGaussianMixtureModel}
\alias{doGaussianMixtureModel}
\title{Do Gaussian mixture modelling for isoform expression differences}
\usage{
doGaussianMixtureModel(dstat, s = NULL, delta, p = NULL, tbreak = NULL,
  useExt = FALSE, delta.thres = 0.75)
}
\arguments{
\item{dstat}{A vector of isoform difference}

\item{s}{A vector of initital values of standard deviation of components}

\item{delta}{A vector of initial values of mean of components}

\item{p}{A vector of initial values of proportion of components}

\item{tbreak}{The number of breaks to create bins}

\item{useExt}{Option of binning tails or not}

\item{delta.thres}{Used for estimated p0}
}
\value{
A list of results and parameters of the mixture model
}
\description{
Do Gaussian mixture modelling for isoform expression differences
}
\examples{
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=logC(isoformDataSample)
#now data is ready
tbreak=round(sqrt(ncol(isoformDataSample)))
x0=isoformDataSample[1,]
x1=isoformDataSample[2,]
deltaVal=x0-x1
s=c(1,1,1);
delta=c(-5,0,5)
res=doGaussianMixtureModel(deltaVal,s=s,delta=delta,tbreak=tbreak)
}

