% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doMixtureModelGene.R
\name{doMixtureModelGene}
\alias{doMixtureModelGene}
\title{Do mixture modelling for all pairs of isoforms of the same single individual gene}
\usage{
doMixtureModelGene(isoMat, geneName = "geneName", usePlot = FALSE,
  useMixModel = TRUE, zero.thres = 0.05, nq.num = 3, tbreak = NULL,
  group.label = NULL)
}
\arguments{
\item{isoMat}{Data matrix contains isoforms of a same individual gene. Each row indicates a isoform, while each column presents a cell}

\item{geneName}{A string contains the name of gene that the isoforms are sharing}

\item{usePlot}{A boolean value to allow plotting data and results. By default usePlot=FALSE}

\item{useMixModel}{A boolean value to allow implementing the mixture model. By default useMixModel=TRUE}

\item{zero.thres}{The threshold to separate components from null component p0. See \code{\link{doMixtureModelIsoformPair}} function}

\item{nq.num}{The maximum components of the mixture model. See \code{\link{doMixtureModelIsoformPair}} function}

\item{tbreak}{The number of breaks to create bins. See \code{\link{doMixtureModelIsoformPair}} function}

\item{group.label}{The group label of cells used for pair plotting}
}
\value{
Lists of results of mixture models for isoform paires
}
\description{
Do mixture modelling for all pairs of isoforms of the same single individual gene
}
\examples{
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=logC(isoformDataSample)
#now data is ready
tbreak=round(sqrt(ncol(isoformDataSample)))
isoMat=isoformDataSample[c(1,2),]
res=doMixtureModelGene(isoMat,geneName="118424",tbreak=tbreak)
}

