% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doMixtureModelIsoformPair.R
\name{doMixtureModelIsoformPair}
\alias{doMixtureModelIsoformPair}
\title{Do mixture modelling for two isoforms}
\usage{
doMixtureModelIsoformPair(x0, x1, nq.num = 3, tbreak = NULL,
  zero.thres = 0.05, min.prop = 0.025)
}
\arguments{
\item{x0}{A vector of data points from isoform 1}

\item{x1}{A vector of data points from isoform 2}

\item{nq.num}{The maximum components of the mixture model}

\item{tbreak}{The number of breaks to create bins}

\item{zero.thres}{The threshold to separate components from null component p0}

\item{min.prop}{Minimum proportion threshold for a component in the mixture model}
}
\value{
A list consisting of list of mixture models (dmix.list), index of the best model (AIC.min.Id), vector of AIC scores (AIC.score), ...
}
\description{
Do mixture modelling for two isoforms
}
\examples{
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=logC(isoformDataSample)
#now data is ready
tbreak=round(sqrt(ncol(isoformDataSample)))
x0=isoformDataSample[1,]
x1=isoformDataSample[2,]
res=doMixtureModelIsoformPair(x0,x1,tbreak=tbreak)
}

