% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMixtureModelMatrix.R
\name{doMixtureModelMatrix}
\alias{doMixtureModelMatrix}
\title{Run mixture models for a data matrix of isoform expression}
\usage{
doMixtureModelMatrix(isoformLevel.data, txdb = NULL, usePlot = FALSE,
  useMixModel = TRUE, zero.thres = 0.05, nq.num = 3, fn.out = NULL,
  tbreak = NULL, group.label = NULL)
}
\arguments{
\item{isoformLevel.data}{A data matrix of isoform expression}

\item{txdb}{A txdb of annotation reference}

\item{usePlot}{A boolean value to allow plotting data and results, usePlot=FALSE (default)}

\item{useMixModel}{A boolean value to allow implementing the mixture model, useMixModel=TRUE (default)}

\item{zero.thres}{The threshold to separate components from null component p0. See \code{\link{doMixtureModelIsoformPair}} function}

\item{nq.num}{The maximum components of the mixture model. See \code{\link{doMixtureModelIsoformPair}} function}

\item{fn.out}{File name of RData to export results}

\item{tbreak}{The number of breaks to create bins. See \code{\link{doMixtureModelIsoformPair}} function}

\item{group.label}{The group label of cells used for pair plotting}
}
\value{
Lists of detected mixture models, AIC scores, number of components and state of models
}
\description{
Run mixture models for a data matrix of isoform expression
}
\examples{
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("AnnotationDbi")
txdb=TxDb.Hsapiens.UCSC.hg19.knownGene
#The txdb also can be imported from a gtf file by using package GenomicFeatures
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=ifelse(isoformDataSample==0,0,log2(isoformDataSample))
#now data is ready
tbreak=round(sqrt(ncol(isoformDataSample)))
res=doMixtureModelMatrix(isoformLevel.data=isoformDataSample,txdb=txdb,tbreak=tbreak)
}
