\name{isoformDataSample}
\alias{isoformDataSample}
\docType{data}
\title{
A sample of read-count expression isoform dataset
}
\description{
A sample of read-count expression isoform dataset
}
\usage{data("isoformDataSample")}
\format{
A data frame with 54 isoforms from 327 cells
  \describe{
    \item{\code{Control_1}}{Cell name}
    \item{\code{Control_2}}{Cell name}
    \item{\code{Control_3}}{Cell name}
    \item{\code{Control_4}}{Cell name}
    \item{\code{Control_5}}{Cell name}
    \item{\code{Control_6}}{Cell name}
    \item{\code{Control_7}}{Cell name}
    \item{\code{Control_8}}{Cell name}
    \item{\code{Control_9}}{Cell name}
    \item{\code{Control_10}}{Cell name}
    \item{\code{Control_11}}{Cell name}
    \item{\code{Control_12}}{Cell name}
    \item{\code{Control_13}}{Cell name}
    \item{\code{Control_14}}{Cell name}
    \item{\code{Control_15}}{Cell name}
    \item{\code{Control_16}}{Cell name}
    \item{\code{Control_17}}{Cell name}
    \item{\code{Control_18}}{Cell name}
    \item{\code{Control_19}}{Cell name}
    \item{\code{Control_20}}{Cell name}
    \item{\code{Control_21}}{Cell name}
    \item{\code{Control_22}}{Cell name}
    \item{\code{Control_23}}{Cell name}
    \item{\code{Control_24}}{Cell name}
    \item{\code{Control_25}}{Cell name}
    \item{\code{Control_26}}{Cell name}
    \item{\code{Control_27}}{Cell name}
    \item{\code{Control_28}}{Cell name}
    \item{\code{Control_29}}{Cell name}
    \item{\code{Control_30}}{Cell name}
    \item{\code{Control_31}}{Cell name}
    \item{\code{Control_32}}{Cell name}
    \item{\code{Control_33}}{Cell name}
    \item{\code{Control_34}}{Cell name}
    \item{\code{Control_35}}{Cell name}
    \item{\code{Control_36}}{Cell name}
    \item{\code{Control_37}}{Cell name}
    \item{\code{Control_38}}{Cell name}
    \item{\code{Control_39}}{Cell name}
    \item{\code{Control_40}}{Cell name}
    \item{\code{Control_41}}{Cell name}
    \item{\code{Control_42}}{Cell name}
    \item{\code{Control_43}}{Cell name}
    \item{\code{Control_44}}{Cell name}
    \item{\code{Control_45}}{Cell name}
    \item{\code{Control_46}}{Cell name}
    \item{\code{Control_47}}{Cell name}
    \item{\code{Control_48}}{Cell name}
    \item{\code{Control_49}}{Cell name}
    \item{\code{Control_50}}{Cell name}
    \item{\code{Control_51}}{Cell name}
    \item{\code{Control_52}}{Cell name}
    \item{\code{Control_53}}{Cell name}
    \item{\code{Control_54}}{Cell name}
    \item{\code{Control_55}}{Cell name}
    \item{\code{Control_56}}{Cell name}
    \item{\code{Control_57}}{Cell name}
    \item{\code{Control_58}}{Cell name}
    \item{\code{Control_59}}{Cell name}
    \item{\code{Control_60}}{Cell name}
    \item{\code{Control_61}}{Cell name}
    \item{\code{Control_62}}{Cell name}
    \item{\code{Control_63}}{Cell name}
    \item{\code{Control_64}}{Cell name}
    \item{\code{Control_65}}{Cell name}
    \item{\code{Control_66}}{Cell name}
    \item{\code{Control_67}}{Cell name}
    \item{\code{Control_68}}{Cell name}
    \item{\code{Control_69}}{Cell name}
    \item{\code{Control_70}}{Cell name}
    \item{\code{Control_71}}{Cell name}
    \item{\code{Control_72}}{Cell name}
    \item{\code{Control_73}}{Cell name}
    \item{\code{Control_74}}{Cell name}
    \item{\code{Control_75}}{Cell name}
    \item{\code{Control_76}}{Cell name}
    \item{\code{Control_77}}{Cell name}
    \item{\code{Control_78}}{Cell name}
    \item{\code{Control_79}}{Cell name}
    \item{\code{Control_80}}{Cell name}
    \item{\code{Control_81}}{Cell name}
    \item{\code{Control_82}}{Cell name}
    \item{\code{Control_83}}{Cell name}
    \item{\code{Control_84}}{Cell name}
    \item{\code{Control_85}}{Cell name}
    \item{\code{Control_86}}{Cell name}
    \item{\code{Control_87}}{Cell name}
    \item{\code{Control_88}}{Cell name}
    \item{\code{Control_89}}{Cell name}
    \item{\code{Control_90}}{Cell name}
    \item{\code{Control_91}}{Cell name}
    \item{\code{Control_92}}{Cell name}
    \item{\code{Control_93}}{Cell name}
    \item{\code{Control_94}}{Cell name}
    \item{\code{Control_95}}{Cell name}
    \item{\code{Control_96}}{Cell name}
    \item{\code{Control_97}}{Cell name}
    \item{\code{Control_98}}{Cell name}
    \item{\code{Control_99}}{Cell name}
    \item{\code{Control_100}}{Cell name}
    \item{\code{Control_101}}{Cell name}
    \item{\code{Control_102}}{Cell name}
    \item{\code{Control_103}}{Cell name}
    \item{\code{Control_104}}{Cell name}
    \item{\code{Control_105}}{Cell name}
    \item{\code{Control_106}}{Cell name}
    \item{\code{Control_107}}{Cell name}
    \item{\code{Control_108}}{Cell name}
    \item{\code{Control_109}}{Cell name}
    \item{\code{Control_110}}{Cell name}
    \item{\code{Control_111}}{Cell name}
    \item{\code{Control_112}}{Cell name}
    \item{\code{Control_113}}{Cell name}
    \item{\code{Control_114}}{Cell name}
    \item{\code{Control_115}}{Cell name}
    \item{\code{Control_116}}{Cell name}
    \item{\code{Control_117}}{Cell name}
    \item{\code{Control_118}}{Cell name}
    \item{\code{Control_119}}{Cell name}
    \item{\code{Control_120}}{Cell name}
    \item{\code{Control_121}}{Cell name}
    \item{\code{Control_122}}{Cell name}
    \item{\code{Control_123}}{Cell name}
    \item{\code{Control_124}}{Cell name}
    \item{\code{Control_125}}{Cell name}
    \item{\code{Control_126}}{Cell name}
    \item{\code{Control_127}}{Cell name}
    \item{\code{Control_128}}{Cell name}
    \item{\code{Control_129}}{Cell name}
    \item{\code{Control_130}}{Cell name}
    \item{\code{Control_131}}{Cell name}
    \item{\code{Control_132}}{Cell name}
    \item{\code{Control_133}}{Cell name}
    \item{\code{Control_134}}{Cell name}
    \item{\code{Control_135}}{Cell name}
    \item{\code{Control_136}}{Cell name}
    \item{\code{Control_137}}{Cell name}
    \item{\code{Control_138}}{Cell name}
    \item{\code{Control_139}}{Cell name}
    \item{\code{Control_140}}{Cell name}
    \item{\code{Control_141}}{Cell name}
    \item{\code{Control_142}}{Cell name}
    \item{\code{Control_143}}{Cell name}
    \item{\code{Control_144}}{Cell name}
    \item{\code{Control_145}}{Cell name}
    \item{\code{Control_146}}{Cell name}
    \item{\code{Control_147}}{Cell name}
    \item{\code{Control_148}}{Cell name}
    \item{\code{Control_149}}{Cell name}
    \item{\code{Control_150}}{Cell name}
    \item{\code{Control_151}}{Cell name}
    \item{\code{Control_152}}{Cell name}
    \item{\code{Control_153}}{Cell name}
    \item{\code{Control_154}}{Cell name}
    \item{\code{Control_155}}{Cell name}
    \item{\code{Control_156}}{Cell name}
    \item{\code{Control_157}}{Cell name}
    \item{\code{Control_158}}{Cell name}
    \item{\code{Control_159}}{Cell name}
    \item{\code{Control_160}}{Cell name}
    \item{\code{Control_161}}{Cell name}
    \item{\code{Control_162}}{Cell name}
    \item{\code{Control_163}}{Cell name}
    \item{\code{Control_164}}{Cell name}
    \item{\code{Control_165}}{Cell name}
    \item{\code{Treated_1}}{Cell name}
    \item{\code{Treated_2}}{Cell name}
    \item{\code{Treated_3}}{Cell name}
    \item{\code{Treated_4}}{Cell name}
    \item{\code{Treated_5}}{Cell name}
    \item{\code{Treated_6}}{Cell name}
    \item{\code{Treated_7}}{Cell name}
    \item{\code{Treated_8}}{Cell name}
    \item{\code{Treated_9}}{Cell name}
    \item{\code{Treated_10}}{Cell name}
    \item{\code{Treated_11}}{Cell name}
    \item{\code{Treated_12}}{Cell name}
    \item{\code{Treated_13}}{Cell name}
    \item{\code{Treated_14}}{Cell name}
    \item{\code{Treated_15}}{Cell name}
    \item{\code{Treated_16}}{Cell name}
    \item{\code{Treated_17}}{Cell name}
    \item{\code{Treated_18}}{Cell name}
    \item{\code{Treated_19}}{Cell name}
    \item{\code{Treated_20}}{Cell name}
    \item{\code{Treated_21}}{Cell name}
    \item{\code{Treated_22}}{Cell name}
    \item{\code{Treated_23}}{Cell name}
    \item{\code{Treated_24}}{Cell name}
    \item{\code{Treated_25}}{Cell name}
    \item{\code{Treated_26}}{Cell name}
    \item{\code{Treated_27}}{Cell name}
    \item{\code{Treated_28}}{Cell name}
    \item{\code{Treated_29}}{Cell name}
    \item{\code{Treated_30}}{Cell name}
    \item{\code{Treated_31}}{Cell name}
    \item{\code{Treated_32}}{Cell name}
    \item{\code{Treated_33}}{Cell name}
    \item{\code{Treated_34}}{Cell name}
    \item{\code{Treated_35}}{Cell name}
    \item{\code{Treated_36}}{Cell name}
    \item{\code{Treated_37}}{Cell name}
    \item{\code{Treated_38}}{Cell name}
    \item{\code{Treated_39}}{Cell name}
    \item{\code{Treated_40}}{Cell name}
    \item{\code{Treated_41}}{Cell name}
    \item{\code{Treated_42}}{Cell name}
    \item{\code{Treated_43}}{Cell name}
    \item{\code{Treated_44}}{Cell name}
    \item{\code{Treated_45}}{Cell name}
    \item{\code{Treated_46}}{Cell name}
    \item{\code{Treated_47}}{Cell name}
    \item{\code{Treated_48}}{Cell name}
    \item{\code{Treated_49}}{Cell name}
    \item{\code{Treated_50}}{Cell name}
    \item{\code{Treated_51}}{Cell name}
    \item{\code{Treated_52}}{Cell name}
    \item{\code{Treated_53}}{Cell name}
    \item{\code{Treated_54}}{Cell name}
    \item{\code{Treated_55}}{Cell name}
    \item{\code{Treated_56}}{Cell name}
    \item{\code{Treated_57}}{Cell name}
    \item{\code{Treated_58}}{Cell name}
    \item{\code{Treated_59}}{Cell name}
    \item{\code{Treated_60}}{Cell name}
    \item{\code{Treated_61}}{Cell name}
    \item{\code{Treated_62}}{Cell name}
    \item{\code{Treated_63}}{Cell name}
    \item{\code{Treated_64}}{Cell name}
    \item{\code{Treated_65}}{Cell name}
    \item{\code{Treated_66}}{Cell name}
    \item{\code{Treated_67}}{Cell name}
    \item{\code{Treated_68}}{Cell name}
    \item{\code{Treated_69}}{Cell name}
    \item{\code{Treated_70}}{Cell name}
    \item{\code{Treated_71}}{Cell name}
    \item{\code{Treated_72}}{Cell name}
    \item{\code{Treated_73}}{Cell name}
    \item{\code{Treated_74}}{Cell name}
    \item{\code{Treated_75}}{Cell name}
    \item{\code{Treated_76}}{Cell name}
    \item{\code{Treated_77}}{Cell name}
    \item{\code{Treated_78}}{Cell name}
    \item{\code{Treated_79}}{Cell name}
    \item{\code{Treated_80}}{Cell name}
    \item{\code{Treated_81}}{Cell name}
    \item{\code{Treated_82}}{Cell name}
    \item{\code{Treated_83}}{Cell name}
    \item{\code{Treated_84}}{Cell name}
    \item{\code{Treated_85}}{Cell name}
    \item{\code{Treated_86}}{Cell name}
    \item{\code{Treated_87}}{Cell name}
    \item{\code{Treated_88}}{Cell name}
    \item{\code{Treated_89}}{Cell name}
    \item{\code{Treated_90}}{Cell name}
    \item{\code{Treated_91}}{Cell name}
    \item{\code{Treated_92}}{Cell name}
    \item{\code{Treated_93}}{Cell name}
    \item{\code{Treated_94}}{Cell name}
    \item{\code{Treated_95}}{Cell name}
    \item{\code{Treated_96}}{Cell name}
    \item{\code{Treated_97}}{Cell name}
    \item{\code{Treated_98}}{Cell name}
    \item{\code{Treated_99}}{Cell name}
    \item{\code{Treated_100}}{Cell name}
    \item{\code{Treated_101}}{Cell name}
    \item{\code{Treated_102}}{Cell name}
    \item{\code{Treated_103}}{Cell name}
    \item{\code{Treated_104}}{Cell name}
    \item{\code{Treated_105}}{Cell name}
    \item{\code{Treated_106}}{Cell name}
    \item{\code{Treated_107}}{Cell name}
    \item{\code{Treated_108}}{Cell name}
    \item{\code{Treated_109}}{Cell name}
    \item{\code{Treated_110}}{Cell name}
    \item{\code{Treated_111}}{Cell name}
    \item{\code{Treated_112}}{Cell name}
    \item{\code{Treated_113}}{Cell name}
    \item{\code{Treated_114}}{Cell name}
    \item{\code{Treated_115}}{Cell name}
    \item{\code{Treated_116}}{Cell name}
    \item{\code{Treated_117}}{Cell name}
    \item{\code{Treated_118}}{Cell name}
    \item{\code{Treated_119}}{Cell name}
    \item{\code{Treated_120}}{Cell name}
    \item{\code{Treated_121}}{Cell name}
    \item{\code{Treated_122}}{Cell name}
    \item{\code{Treated_123}}{Cell name}
    \item{\code{Treated_124}}{Cell name}
    \item{\code{Treated_125}}{Cell name}
    \item{\code{Treated_126}}{Cell name}
    \item{\code{Treated_127}}{Cell name}
    \item{\code{Treated_128}}{Cell name}
    \item{\code{Treated_129}}{Cell name}
    \item{\code{Treated_130}}{Cell name}
    \item{\code{Treated_131}}{Cell name}
    \item{\code{Treated_132}}{Cell name}
    \item{\code{Treated_133}}{Cell name}
    \item{\code{Treated_134}}{Cell name}
    \item{\code{Treated_135}}{Cell name}
    \item{\code{Treated_136}}{Cell name}
    \item{\code{Treated_137}}{Cell name}
    \item{\code{Treated_138}}{Cell name}
    \item{\code{Treated_139}}{Cell name}
    \item{\code{Treated_140}}{Cell name}
    \item{\code{Treated_141}}{Cell name}
    \item{\code{Treated_142}}{Cell name}
    \item{\code{Treated_143}}{Cell name}
    \item{\code{Treated_144}}{Cell name}
    \item{\code{Treated_145}}{Cell name}
    \item{\code{Treated_146}}{Cell name}
    \item{\code{Treated_147}}{Cell name}
    \item{\code{Treated_148}}{Cell name}
    \item{\code{Treated_149}}{Cell name}
    \item{\code{Treated_150}}{Cell name}
    \item{\code{Treated_151}}{Cell name}
    \item{\code{Treated_152}}{Cell name}
    \item{\code{Treated_153}}{Cell name}
    \item{\code{Treated_154}}{Cell name}
    \item{\code{Treated_155}}{Cell name}
    \item{\code{Treated_156}}{Cell name}
    \item{\code{Treated_157}}{Cell name}
    \item{\code{Treated_158}}{Cell name}
    \item{\code{Treated_159}}{Cell name}
    \item{\code{Treated_160}}{Cell name}
    \item{\code{Treated_161}}{Cell name}
    \item{\code{Treated_162}}{Cell name}
  }
}
\examples{
data(isoformDataSample)
}
\keyword{datasets}
