\name{ISOP-package}
\alias{ISOP-package}
\alias{ISOP}
\docType{package}
\title{
Isoform-level Expression Patterns in Single-cell RNA-sequencing Data
}
\description{
Isoform-level Expression Patterns in Single-cell RNA-sequencing Data
}
\details{
\tabular{ll}{
Package: \tab ISOP\cr
Type: \tab Package\cr
Version: \tab 0.99.1\cr
Date: \tab 2017-02-24\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Trung Nghia Vu, Yudi Pawitan and Mattias Rantalainen

Maintainer: Trung Nghia Vu <nghiavtr@gmail.com>

}
\examples{ 
# See vignette
vignette("ISOP")
}

\keyword{ package }


