% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logC.R
\name{logC}
\alias{logC}
\title{Get log scale of read counr vector or matrix}
\usage{
logC(x, type = "log2", log.zero = 0)
}
\arguments{
\item{x}{A vector or matrix of read counts}

\item{type}{Type of transformation log2, log10 or log}

\item{log.zero}{Value that zero values of read counts will become in log scale (=0 by default)}
}
\value{
Data values after log scaling
}
\description{
Return log2, log10 or log values of expressed values while assign specific value (log.zero) to zero values
}
\examples{
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=logC(isoformDataSample)
#now data is ready
}

