% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validateIsoformPair.R
\name{validateIsoformPair}
\alias{validateIsoformPair}
\title{Validate the randomness of isoform pair via permutation}
\usage{
validateIsoformPair(iso.pair.names = NULL, isoformLevel.data,
  geneLevel.data = NULL, per.num = 10000, tbreak = 10, esp.val = 0.5,
  useExt = FALSE, useParallel = FALSE)
}
\arguments{
\item{iso.pair.names}{A vector of names of isoform pairs in the format: gene.isoform1.isoform2}

\item{isoformLevel.data}{A data matrix of isoform expression}

\item{geneLevel.data}{A data matrix of gene expression}

\item{per.num}{A number of permutation}

\item{tbreak}{The number of breaks to create bins}

\item{esp.val}{A small value added to denominator of X2 formula}

\item{useExt}{A boolean value to allow the left-most and right-most bins extending to infinity}

\item{useParallel}{An option for using parallel (=TRUE)}
}
\value{
A list containing results of X2 tests and p-values
}
\description{
Validate the randomness of isoform pair via permutation
}
\examples{
library("TxDb.Hsapiens.UCSC.hg19.knownGene")
library("AnnotationDbi")
txdb=TxDb.Hsapiens.UCSC.hg19.knownGene
#The txdb also can be imported from a gtf file by using package GenomicFeatures
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=logC(isoformDataSample)
#now data is ready
tbreak=round(sqrt(ncol(isoformDataSample)))
model.res=doMixtureModelMatrix(isoformLevel.data=isoformDataSample,txdb=txdb,tbreak=tbreak)
#library(doParallel)
#registerDoParallel(cores=4)
set.seed(2015)
iso.pair.names=names(model.res$nq.list)
res=validateIsoformPair(iso.pair.names=iso.pair.names,
	   isoformLevel.data=isoformDataSample,per.num=100,tbreak=tbreak,useParallel=FALSE)
PVAL.X2.list=unlist(res[names(res)=="pval"])
hist(PVAL.X2.list)
fdr.val=p.adjust(PVAL.X2.list,method="BH")
length(which(fdr.val <= 0.05))
}

