% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPairFeatures.R
\name{plotPairFeatures}
\alias{plotPairFeatures}
\title{Plot two isoforms by pair lines}
\usage{
plotPairFeatures(x0, x1, group.label = NULL, group.color = c("red", "blue",
  "green", "pink"))
}
\arguments{
\item{x0}{A vector of data points of the first isoforms}

\item{x1}{A vector of data points of the second isoforms}

\item{group.label}{A vector of group label of cells}

\item{group.color}{A vector of colors for the cell groups}
}
\value{
None
}
\description{
Plot two isoforms by pair lines
}
\examples{
data(isoformDataSample)
#preprocessing
isoformDataSample=ifelse(isoformDataSample <= 3,0,isoformDataSample)
isoformDataSample=isoformDataSample[which(rowSums(isoformDataSample)>0),]
#tranform read count dataset to log scale
isoformDataSample=logC(isoformDataSample)
#now data is ready
x0=isoformDataSample[1,]
x1=isoformDataSample[2,]
plotPairFeatures(x0,x0)
}

